Phase4_Backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0              # async PostgreSQL driver (used by SQLAlchemy async)
psycopg2-binary==2.9.9       # sync driver (used by Alembic migrations)

# ── Auth ───────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT encode/decode
passlib[bcrypt]==1.7.4             # password hashing
httpx==0.27.0                      # async HTTP client (Google OAuth token exchange)
itsdangerous==2.2.0                # secure token signing (email verification)

# ── Task Queue ─────────────────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.4                       # Redis client (Celery broker + result backend)

# ── File Handling ──────────────────────────────────────────────────────────────
python-multipart==0.0.9            # multipart/form-data uploads (FastAPI)
pypdf==4.2.0                       # PDF text extraction
python-docx==1.1.2                 # DOCX text extraction
aiofiles==23.2.1                   # async file I/O

# ── File Storage ───────────────────────────────────────────────────────────────
boto3==1.34.101                    # AWS S3 / MinIO client

# ── Data Validation ────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1           # settings management via .env

# ── Email ──────────────────────────────────────────────────────────────────────
fastapi-mail==1.4.1

# ── Report Generation ──────────────────────────────────────────────────────────
weasyprint==62.3                   # PDF report export

# ── Rate Limiting ──────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Logging ────────────────────────────────────────────────────────────────────
python-json-logger==2.0.7

# ── Dev / Testing ──────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0                      # also used for test client
python-dotenv==1.0.1